data_pipeline/data_reader/requirements.txt
boto3==1.26.41
pandas==1.5.2
psycopg2-binary
user_agents